requirements.txt
django-bootstrap-datepicker-plus
django-tables2
django-crispy-forms
bootstrap411

